update data fetch manager to add city_visitor

+++ src/database_manager.py
@@ -18,6 +18,8 @@ class City(Base):
     population = Column(String)
     size = Column(String)
     year_reported = Column(String)
+    city_visitor = Column(String)
+    city_visitor_reported_year = Column(String)
 
     def __init__(self, city_infos):
         super().__init__()
@@ -34,6 +36,14 @@ class City(Base):
         key ="population_as_of"
         if key in city_infos:
             self.year_reported = city_infos[key]
+
+        key ="city_visitor"
+        if key in city_infos:
+            self.city_visitor = city_infos[key]
+
+        key = "city_visitor_reported_year"
+        if key in city_infos:
+            self.city_visitor_reported_year = city_infos[key]
 
     def is_valid(self):
         # todo: modify this to be depend on user needs
@@ -137,11 +147,17 @@ class DatabaseManager():
 
     def __init__(self):
         self.database_type = {"postgres": "postgres"}
+        self.initialized = False
         return
 
     def init(self, config ):
+        if self.initialized:
+            return
+
         if config.database_type not in self.database_type:
             raise AssertionError("currently " + config.database_type + " database does not supported by the system!")
+
+        self.initialized = True
 
         # build connection string
         db_string = config.database_type + "://" + config.database_user_name + ":" + \
@@ -159,10 +175,13 @@ class DatabaseManager():
             self.drop_table(Museum.__tablename__)
             self.drop_table(City.__tablename__)
 
-
         self.metadata.create_all(self.db_engine)
 
     def save(self, **data):
+        if not self.initialized:
+            # log database is not initialized
+            return
+
         if not isinstance(data, dict):
             raise AssertionError("data should be provided as a dict")
         if len(data)==0:
@@ -227,11 +246,21 @@ class DatabaseManager():
             df_parsed_table=df_parsed_table.append(pd.Series(values, index=keys), ignore_index=True)
         return df_parsed_table
 
+    def objects_valid(self, objects):
+        for object in objects:
+            if not object.is_valid():
+                return False
+        return True
+
     def query_to_list(self, rset):
         result = []
         for row in rset:
             row_data = {}
+            if not self.objects_valid(row):
+                continue
+
             for object in row:
+                object.is_valid()
                 object_in_dict = self.object_as_dict(object)
                 self.keys_swap("name", object.__tablename__, object_in_dict)
                 row_data.update(object_in_dict)
@@ -240,6 +269,10 @@ class DatabaseManager():
         return result
 
     def load(self):
+        if not self.initialized:
+            # log database is not initialized
+            return
+
         try:
             query_result = self.session.query(City, Museum).filter( Museum.city_id == City.id).all()
             return self.query_to_df(query_result)
@@ -248,11 +281,19 @@ class DatabaseManager():
             raise E.args[0]
 
     def delete_all_data(self):
+        if not self.initialized:
+            # log database is not initialized
+            return
+
         self.session.query(Museum).delete()
         self.session.query(City).delete()
         self.session.commit()
 
     def drop_table(self,table_name):
+        if not self.initialized:
+            # log database is not initialized
+            return
+
         table = self.metadata.tables.get(table_name)
         if table is not None:
             # logging.info(f'Deleting {table_name} table')
